ARM: config: Add BCM53xx defconfig
This adds a BCM530x ARM-based Broadcom default config, typically
for use with NorthStar systems. This is needed because it took
me quite a lot of time to figure out all the different drivers and
intricacies that needed to be enabled to bring a typical specimen
from these systems up.

By simply using this it is easy to configure and boot a mainline
Linux kernel on these systems.

Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
---
I suppose Florian can merge this, I can also send it directly to
the SOC tree if preferred.