net: ethernet: cortina: Restore TSO support
An earlier commit deleted the TSO support in the Cortina Gemini
driver because the driver was confusing gso_size and MTU,
probably because what the Linux kernel calls "gso_size" was
called "MTU" in the datasheet.

Restore the functionality properly reading the gso_size from
the skbuff.

Tested with iperf3, running a server on a different machine
and client on the device with the cortina gemini ethernet:

Connecting to host 192.168.1.2, port 5201
60008000.ethernet-port eth0: segment offloading mss = 05a8 len=1c8a
60008000.ethernet-port eth0: segment offloading mss = 05a8 len=1c8a
60008000.ethernet-port eth0: segment offloading mss = 05a8 len=27da
60008000.ethernet-port eth0: segment offloading mss = 05a8 len=0b92
60008000.ethernet-port eth0: segment offloading mss = 05a8 len=2bda
(...)

It also performs well: ~268 MBit/s.

Fixes: ac631873c9e7 ("net: ethernet: cortina: Drop TSO support")
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>