sched/eevdf: Augment comments to account for reality
The references to "CFS" is a bit misleading these days since
the scheduling principe is EEVDF.

Rewrite the top level comment, and trim other comments and
kerneldoc to implicitly refer to the scheduling implemented
in this file, or just "the fair scheduler".

Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
---
This isn't fixing everything and isn't changing any kernel
symbols containing CFS, as I think that would be churny.